добавлено подтверждение при удалении записей

+++ frontend/src/views/Department/DepartmentList/DepartmentList.tsx
@@ -7,6 +7,8 @@ import CircularIndeterminate from "components/Loader/Loader";
 import { DefaultToolbar} from 'components';
 import {IStateInterface} from "redux/rootReducer";
 import {deleteDepartment, loadDepartments} from "redux/actions/departmentAction";
+import {useConfirm} from "material-ui-confirm";
+import {DIALOG_ASK_DELETE, DIALOG_NO, DIALOG_TYPE_CONFIRM, DIALOG_YES} from "../../../utils/AppConst";
 
 
 const useStyles = makeStyles(theme => ({
@@ -19,9 +21,10 @@ const useStyles = makeStyles(theme => ({
 }));
 
 const DepartmentList = () => {
-    const classes = useStyles();
-    const history = useHistory();
-    const dispatch = useDispatch();
+    const classes = useStyles()
+    const history = useHistory()
+    const dispatch = useDispatch()
+    const confirm = useConfirm()
 
     const items = useSelector((state: IStateInterface) => state.department.items);
     const isLoading = useSelector((state: IStateInterface) => state.department.isLoading);
@@ -36,10 +39,23 @@ const DepartmentList = () => {
         dispatch(loadDepartments(findText))
     }
 
+    /**
+     * Удалить записи
+     */
     function onDeleteHandle() {
-        selected.forEach(async (item, i, selected) => {
-            dispatch(deleteDepartment(item))
-        });
+        confirm(
+            {
+                'title': DIALOG_TYPE_CONFIRM,
+                description: DIALOG_ASK_DELETE,
+                confirmationText: DIALOG_YES,
+                cancellationText: DIALOG_NO
+            }
+        ).then(() =>
+            selected.forEach(async (item) => {
+                dispatch(deleteDepartment(item))
+            })
+        )
+
     }
 
     function onClickTableItem(id: number){
